pyEngine/requirements.txt
rembg>=2.0.50
pillow>=10.0.0
onnxruntime>=1.16.0
google-generativeai>=0.8.4
